[Language] - rename modules and classes to reflect specific concerns
Most importantly, the package is now 'localization' (its overarching
concern) to distinguish it from 'localizer' (the name of one class)
