+++ localization/src/main/AndroidManifest.xml
@@ -1,2 +1,2 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="edu.artic.localizer" />
+    package="edu.artic.localization" />
